+++ trunk/interfaces/src/main/java/net/sf/otrcutmp4/interfaces/facade/OtrMediacenterFacade.java
@@ -16,8 +16,8 @@ public interface OtrMediacenterFacade<MOVIE extends Movie<COVER,STORAGE>,SERIES 
 	MOVIE fMovie(Class<MOVIE> type, String name, int year) throws UtilsNotFoundException;
 	
 	SERIES fSeries(Class<SERIES> type, String name) throws UtilsNotFoundException;
-	SEASON fSeason(Class<SEASON> type, SERIES series, int nr) throws UtilsNotFoundException;
-	EPISODE fEpisode(Class<EPISODE> type, SEASON season, int nr) throws UtilsNotFoundException;
+	SEASON fSeason(Class<SEASON> type, SERIES series, long nr) throws UtilsNotFoundException;
+	EPISODE fEpisode(Class<EPISODE> type, SEASON season, long nr) throws UtilsNotFoundException;
 	
 	EPISODE fcEpisode(Class<SERIES> clSeries, Class<SEASON> clSeason, Class<EPISODE> clEpisode, Class<COVER> clCover, net.sf.otrcutmp4.model.xml.series.Episode episode);
 	SEASON fcSeason(Class<SEASON> clSeason, SERIES series, net.sf.otrcutmp4.model.xml.series.Season season);
